Base file to just get audio played.
Didn't really write much in here, distortion is not working on this level for one reason or another I am unsure. Yet if you put actual code in this it works visually. I believe the issue is it goes in relation to the xy moving forward and I don't have proper calculations for the Fast Fourier Transformation currently of my audio as well as sample sizing.