Finalized Post Processor
The bound audio is really effected with this pass. I think for some reason it takes information from this pass in particular alongside what was already bound to our audio file for the distortion to take hold.
Otherwords: Audio pass holds the sample while this pass feeds the tone, coincidentally the audio pass gets this pass' distortion on the audio regardless of which end of the pipeline this is in. Not sure why.